Fix for brew readline built x86_64 only
The keg'd GNU readline library is not universal x86_64/i386 but instead
is x86_64 only.  The build of the Perl Term::ReadLine::Gnu thus fails by
default because it tries to be Universal.

So make an x86_64-only readline.

This may cause compatibility problems if this formula is run on very old
hardware which is i386-only.  If that's reported as an actual issue in
practice, then I'll look into what it takes to look at the library and
dynamically figure out the arch build env controls needed.