Handle Brew no longer installing Perl modules for us
sieve-connect to revision 1 (still 0.89 version).

We can no longer use `depends_on "Foo" => :perl` to have the perl module
Foo.  Instead, we have to vendor everything.  So do so.

Also switch options to with/without style and fix most, but not all, of
the issues identified by `brew audit --strict --online sieve-connect`.
This changes the options, but the old names are kept as deprecated
aliases.