dumpasn1: .cfg modified, switch to https
Upstream has made http: into redirects for https: (whereas when I last
checked https: was not available) so use https immediately.

Bump to revision 1 to pick up a .cfg change upstream, fixing an OID arc
for `cryptoProDigestD`; it had the same arc as `cryptoProDigestC` and
that was repaired without changing the internal datestamp.  I picked the
date returned by the webserver in the last-modified timestamp.